Convert all surrogates to web accessible resources
