+++ src/data/web_accessible_resources/amazon_apstag.js
@@ -1,22 +1,19 @@
 // https://github.com/gorhill/uBlock/blob/4b95420e5912cb2759da77dbb3d0d64095021c13/src/web_accessible_resources/amazon_apstag.js
-
-(function () {
-
-'use strict';
-const w = window;
-const noopfn = function () {
-  ; // eslint-disable-line no-extra-semi
-}.bind();
-const apstag = {
-  fetchBids: function (a, b) {
-    if (b instanceof Function) {
-      b([]);
-    }
-  },
-  init: noopfn,
-  setDisplayBids: noopfn,
-  targetingKeys: noopfn,
-};
-w.apstag = apstag;
-
-}());
+(function() {
+    'use strict';
+    const w = window;
+    const noopfn = function() {
+        ; // jshint ignore:line
+    }.bind();
+    const apstag = {
+        fetchBids: function(a, b) {
+            if ( b instanceof Function ) {
+                b([]);
+            }
+        },
+        init: noopfn,
+        setDisplayBids: noopfn,
+        targetingKeys: noopfn,
+    };
+    w.apstag = apstag;
+})();
