+++ src/data/web_accessible_resources/youneeq.js
@@ -1,0 +1,14 @@
+(function () {
+    var noopfn = function() {
+        ;
+    };
+    function YqClass() {
+        this.observe = noopfn;
+        this.observeMin = noopfn;
+        this.scroll_event = noopfn;
+        this.onready = noopfn;
+        this.yq_panel_click = noopfn;
+        this.titleTrim = noopfn;
+    }
+    window.Yq || (window.Yq = new YqClass); // eslint-disable-line no-unused-expressions
+}());
