+++ src/js/constants.js
@@ -30,7 +30,6 @@ let exports = {
   USER_COOKIEBLOCK: "user_cookieblock",
 
   // URLS
-  WEBEXT_SCHEME: chrome.runtime.getURL(''),
   CNAME_DOMAINS_LOCAL_URL: chrome.runtime.getURL('data/cname_domains.json'),
   DNT_POLICIES_URL: "https://www.eff.org/files/dnt-policies.json",
   DNT_POLICIES_LOCAL_URL: chrome.runtime.getURL('data/dnt-policies.json'),
